+++ src/lib/hooks/tests/server_hooks_unittest.cc
@@ -1,4 +1,4 @@
-// Copyright (C) 2013  Internet Systems Consortium, Inc. ("ISC")
+// Copyright (C) 2013, 2015  Internet Systems Consortium, Inc. ("ISC")
 //
 // Permission to use, copy, modify, and/or distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -88,7 +88,7 @@ TEST(ServerHooksTest, GetHookNames) {
     expected_names.push_back("beta");
     expected_names.push_back("gamma");
     expected_names.push_back("delta");
-    for (int i = 0; i < expected_names.size(); ++i) {
+    for (size_t i = 0; i < expected_names.size(); ++i) {
         hooks.registerHook(expected_names[i].c_str());
     };
